synthetic_distillation/requirements.txt
hydra-core>=1.3.0
hydra-colorlog>=1.2.0
torch
tensorboard
gymnasium[mujoco]
numpy
